Add files via upload
The Employee Management System (EMS) is a software project aimed at automating and streamlining the management of employee records, payroll, and attendance tracking. The system is designed to be user-friendly, efficient, and customizable to suit the needs of small to medium-sized businesses. The project is developed using the C programming language and utilizes a relational database management system to store and manage employee data. The EMS has several features including a login system, employee data management, payroll processing and leave management.